serializer changes extractet from simple-deserializer branch
